LPS-85088 Expand UploadServletRequestTest into individual test classes
